+++ src/components/Button/Button.module.css
@@ -1,16 +1,62 @@
-.btn {
-  padding: 0.5rem 1rem;
-  border-radius: 0.2rem;
-  color: #ffffff;
-  background-color: #3369ff;
-  box-shadow: inset 0 -0.6rem 1rem -0.35rem rgba(0, 0, 0, 0.17), inset 0 0.6rem 2rem -0.3rem rgba(255, 255, 255, 0.15),
-    inset 0 0 0rem 0.05rem rgba(255, 255, 255, 0.12);
+.Button {
+  font-size: var(--button-text);
+  height: 3.25rem;
+  padding: 0.625rem 1.6rem;
+  border-radius: 3.125rem;
+  color: var(--common-white);
+  transition-duration: var(--default-transition);
 }
 
-.btn:hover {
-  opacity: 0.9;
+.ButtonOutlined {
+  composes: Button;
+  border: 0.1875rem solid var(--secondary-main);
+}
+.ButtonNormal {
+  composes: Button;
+  background-color: var(--secondary-main);
 }
 
-.btn:active {
-  box-shadow: inset 0 0.6rem 2rem -0.3rem rgba(0, 0, 0, 0.15), inset 0 0 0rem 0.05rem rgba(255, 255, 255, 0.12);
+.ButtonNextQuestion {
+  composes: Button;
+  background-color: var(--secondary-dark);
+}
+
+.ButtonGameType {
+  composes: Button;
+  width: 100%;
+  height: 4.375rem;
+  border: 6px solid var(--primary-light);
+}
+
+.Button:hover {
+  background-color: var(--secondary-dark);
+  border-color: transparent;
+}
+
+.Button:disabled {
+  color: rgba(255, 255, 255, 0.38);
+  background-color: rgba(255, 255, 255, 0.5);
+  border-color: transparent;
+  cursor: initial;
+}
+
+.ButtonNextQuestion:hover {
+  background-color: var(--secondary-main);
+  border-color: transparent;
+}
+
+.ButtonNextQuestion:active {
+  background-color: var(--secondary-light);
+  color: var(--secondary-dark);
+  border-color: transparent;
+}
+
+.ButtonGameType:hover {
+  background-color: var(--primary-main);
+  border: 6px solid var(--primary-main);
+}
+
+.ButtonGameType:active {
+  background-color: var(--primary-light);
+  border-color: transparent;
 }
