--- src/components/GameTypeButton/index.js
@@ -1,1 +1,0 @@
-export * from './GameTypeButton';
