--- src/components/MenuButton/index.js
@@ -1,1 +1,0 @@
-export * from './MenuButton';
