+++ src/components/index.js
@@ -1,4 +1,3 @@
 export * from './Button';
 export * from './GameTypes';
-export * from './MenuButton';
 export * from './RulesTextBox';
